browserify: basic build.
The core browserify build is now working.  Try it with:

node_modules/.bin/beefy test/spec/0indexing-spec.js -- -i WNdb -i lapack

(The two -i options are to ignore unused modules
that are conditionally require()'d in NaturalNode,
but not declared as dependnecies and thus not
installed.)

Still something not quite right with
initializing the database, but getting close!