This commit is a rejig of the build process for "webification"
Now the general strategy is that the lib is written in the commons.js
style (using require, module.exports, etc), and used in an uncompiled
state for node applications. It is only compiled for use on the web,
and only webpack is used for this. At the moment (babel) transpilation for
backwards compatability is stripped out, although this should probably
be added in again at some point.